[ADD] 유효성 검사 추가 등
